requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
numpy
pydicom
Pillow
python-multipart